# cardano-rosetta-server/src/server/openApi.json
{
  "openapi": "3.0.2",
  "info": {
    "version": "1.4.1",
    "title": "Rosetta",
    "description": "Build Once. Integrate Your Blockchain Everywhere.",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "paths": {
...
    "/block/transaction": {
      "post": {
        "summary": "Get a Block Transaction",
        "description": "Get a transaction in a block by its Transaction Identifier. This endpoint should only be used when querying a node for a block does not return all transactions contained within it. All transactions returned by this endpoint must be appended to any transactions returned by the /block method by consumers of this data. Fetching a transaction by hash is considered an Explorer Method (which is classified under the Future Work section). Calling this endpoint requires reference to a BlockIdentifier because transaction parsing can change depending on which block contains the transaction. For example, in Bitcoin it is necessary to know which block contains a transaction to determine the destination of fee payments. Without specifying a block identifier, the node would have to infer which block to use (which could change during a re-org). Implementations that require fetching previous transactions to populate the response (ex: Previous UTXOs in Bitcoin) may find it useful to run a cache within the Rosetta server in the /data directory (on a path that does not conflict with the node).",
        "operationId": "blockTransaction",
        "tags": ["Block"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockTransactionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockTransactionResponse"
                }
              }
            }
          },
          "500": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "NetworkIdentifier": {
        "description": "The network_identifier specifies which network a particular object is associated with.",
        "type": "object",
        "required": ["blockchain", "network"],
        "properties": {
          "blockchain": {
            "type": "string",
            "example": "bitcoin"
          },
          "network": {
            "description": "If a blockchain has a specific chain-id or network identifier, it should go in this field. It is up to the client to determine which network-specific identifier is mainnet or testnet.",
            "type": "string",
            "example": "mainnet"
          },
          "sub_network_identifier": {
            "$ref": "#/components/schemas/SubNetworkIdentifier"
          }
        }
      },
      "SubNetworkIdentifier": {
        "description": "In blockchains with sharded state, the SubNetworkIdentifier is required to query some object on a specific shard. This identifier is optional for all non-sharded blockchains.",
        "type": "object",
        "required": ["network"],
        "properties": {
          "network": {
            "type": "string",
            "example": "shard 1"
          },
          "metadata": {
            "type": "object",
            "example": {
              "producer": "0x52bc44d5378309ee2abf1539bf71de1b7d7be3b5"
            }
          }
        }
      },
      "BlockIdentifier": {
        "description": "The block_identifier uniquely identifies a block in a particular network.",
        "type": "object",
        "required": ["index", "hash"],
        "properties": {
          "index": {
            "description": "This is also known as the block height.",
            "type": "integer",
            "format": "int64",
            "example": 1123941
          },
          "hash": {
            "type": "string",
            "example": "0x1f2cc6c5027d2f201a5453ad1119574d2aed23a392654742ac3c78783c071f85"
          }
        }
      },
...
      "TransactionIdentifier": {
        "description": "The transaction_identifier uniquely identifies a transaction in a particular network and block or in the mempool.",
        "type": "object",
        "required": ["hash"],
        "properties": {
          "hash": {
            "description": "Any transactions that are attributable only to a block (ex: a block event) should use the hash of the block as the identifier.",
            "type": "string",
            "example": "0x2f23fd8cca835af21f3ac375bac601f97ead75f2e79143bdf71fe2c4be043e8f"
          }
        }
      },
      "OperationIdentifier": {
        "description": "The operation_identifier uniquely identifies an operation within a transaction.",
        "type": "object",
        "required": ["index"],
        "properties": {
          "index": {
            "description": "The operation index is used to ensure each operation has a unique identifier within a transaction. This index is only relative to the transaction and NOT GLOBAL. The operations in each transaction should start from index 0. To clarify, there may not be any notion of an operation index in the blockchain being described.",
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "example": 1
          },
          "network_index": {
            "description": "Some blockchains specify an operation index that is essential for client use. For example, Bitcoin uses a network_index to identify which UTXO was used in a transaction. network_index should not be populated if there is no notion of an operation index in a blockchain (typically most account-based blockchains).",
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "example": 0
          }
        }
      },
      "AccountIdentifier": {
        "description": "The account_identifier uniquely identifies an account within a network. All fields in the account_identifier are utilized to determine this uniqueness (including the metadata field, if populated).",
        "type": "object",
        "required": ["address"],
        "properties": {
          "address": {
            "description": "The address may be a cryptographic public key (or some encoding of it) or a provided username.",
            "type": "string",
            "example": "0x3a065000ab4183c6bf581dc1e55a605455fc6d61"
          },
          "sub_account": {
            "$ref": "#/components/schemas/SubAccountIdentifier"
          },
          "metadata": {
            "description": "Blockchains that utilize a username model (where the address is not a derivative of a cryptographic public key) should specify the public key(s) owned by the address in metadata.",
            "type": "object"
          }
        }
      },
      "SubAccountIdentifier": {
        "description": "An account may have state specific to a contract address (ERC-20 token) and/or a stake (delegated balance). The sub_account_identifier should specify which state (if applicable) an account instantiation refers to.",
        "type": "object",
        "required": ["address"],
        "properties": {
          "address": {
            "description": "The SubAccount address may be a cryptographic value or some other identifier (ex: bonded) that uniquely specifies a SubAccount.",
            "type": "string",
            "example": "0x6b175474e89094c44da98b954eedeac495271d0f"
          },
          "metadata": {
            "description": "If the SubAccount address is not sufficient to uniquely specify a SubAccount, any other identifying information can be stored here. It is important to note that two SubAccounts with identical addresses but differing metadata will not be considered equal by clients.",
            "type": "object"
          }
        }
      },
...
      "Transaction": {
        "description": "Transactions contain an array of Operations that are attributable to the same TransactionIdentifier.",
        "type": "object",
        "required": ["transaction_identifier", "operations"],
        "properties": {
          "transaction_identifier": {
            "$ref": "#/components/schemas/TransactionIdentifier"
          },
          "operations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Operation"
            }
          },
          "metadata": {
            "description": "Transactions that are related to other transactions (like a cross-shard transaction) should include the tranaction_identifier of these transactions in the metadata.",
            "type": "object",
            "example": {
              "size": 12378,
              "lockTime": 1582272577
            }
          }
        }
      },
      "Operation": {
        "description": "Operations contain all balance-changing information within a transaction. They are always one-sided (only affect 1 AccountIdentifier) and can succeed or fail independently from a Transaction.",
        "type": "object",
        "required": ["operation_identifier", "type", "status"],
        "properties": {
          "operation_identifier": {
            "$ref": "#/components/schemas/OperationIdentifier"
          },
          "related_operations": {
            "description": "Restrict referenced related_operations to identifier indexes < the current operation_identifier.index. This ensures there exists a clear DAG-structure of relations. Since operations are one-sided, one could imagine relating operations in a single transfer or linking operations in a call tree.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OperationIdentifier"
            },
            "example": [
              {
                "index": 0,
                "operation_identifier": {
                  "index": 0
                }
              }
            ]
          },
          "type": {
            "description": "The network-specific type of the operation. Ensure that any type that can be returned here is also specified in the NetworkStatus. This can be very useful to downstream consumers that parse all block data.",
            "type": "string",
            "example": "Transfer"
          },
          "status": {
            "description": "The network-specific status of the operation. Status is not defined on the transaction object because blockchains with smart contracts may have transactions that partially apply. Blockchains with atomic transactions (all operations succeed or all operations fail) will have the same status for each operation.",
            "type": "string",
            "example": "Reverted"
          },
          "account": {
            "$ref": "#/components/schemas/AccountIdentifier"
          },
          "amount": {
            "$ref": "#/components/schemas/Amount"
          },
          "coin_change": {
            "$ref": "#/components/schemas/CoinChange"
          },
          "metadata": {
            "type": "object",
            "example": {
              "asm": "304502201fd8abb11443f8b1b9a04e0495e0543d05611473a790c8939f089d073f90509a022100f4677825136605d732e2126d09a2d38c20c75946cd9fc239c0497e84c634e3dd01 03301a8259a12e35694cc22ebc45fee635f4993064190f6ce96e7fb19a03bb6be2",
              "hex": "48304502201fd8abb11443f8b1b9a04e0495e0543d05611473a790c8939f089d073f90509a022100f4677825136605d732e2126d09a2d38c20c75946cd9fc239c0497e84c634e3dd012103301a8259a12e35694cc22ebc45fee635f4993064190f6ce96e7fb19a03bb6be2"
            }
          }
        }
      },
      "Amount": {
        "description": "Amount is some Value of a Currency. It is considered invalid to specify a Value without a Currency.",
        "type": "object",
        "required": ["value", "currency"],
        "properties": {
          "value": {
            "description": "Value of the transaction in atomic units represented as an arbitrary-sized signed integer. For example, 1 BTC would be represented by a value of 100000000.",
            "type": "string",
            "example": "1238089899992"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          },
          "metadata": {
            "type": "object"
          }
        }
      },
...
      "Currency": {
        "description": "Currency is composed of a canonical Symbol and Decimals. This Decimals value is used to convert an Amount.Value from atomic units (Satoshis) to standard units (Bitcoins).",
        "type": "object",
        "required": ["symbol", "decimals"],
        "properties": {
          "symbol": {
            "description": "Canonical symbol associated with a currency.",
            "type": "string",
            "example": "BTC"
          },
          "decimals": {
            "description": "Number of decimal places in the standard unit representation of the amount. For example, BTC has 8 decimals. Note that it is not possible to represent the value of some currency in atomic units that is not base 10.",
            "type": "integer",
            "format": "int32",
            "minimum": 0,
            "example": 8
          },
          "metadata": {
            "description": "Any additional information related to the currency itself. For example, it would be useful to populate this object with the contract address of an ERC-20 token.",
            "type": "object",
            "example": {
              "Issuer": "Satoshi"
            }
          }
        }
      },
...
      "CoinAction": {
        "description": "CoinActions are different state changes that a Coin can undergo. When a Coin is created, it is coin_created. When a Coin is spent, it is coin_spent. It is assumed that a single Coin cannot be created or spent more than once.",
        "type": "string",
        "enum": ["coin_created", "coin_spent"]
      },
      "CoinIdentifier": {
        "description": "CoinIdentifier uniquely identifies a Coin.",
        "type": "object",
        "required": ["identifier"],
        "properties": {
          "identifier": {
            "description": "Identifier should be populated with a globally unique identifier of a Coin. In Bitcoin, this identifier would be transaction_hash:index.",
            "type": "string",
            "example": "0x2f23fd8cca835af21f3ac375bac601f97ead75f2e79143bdf71fe2c4be043e8f:1"
          }
        }
      },
      "CoinChange": {
        "description": "CoinChange is used to represent a change in state of a some coin identified by a coin_identifier. This object is part of the Operation model and must be populated for UTXO-based blockchains. Coincidentally, this abstraction of UTXOs allows for supporting both account-based transfers and UTXO-based transfers on the same blockchain (when a transfer is account-based, don't populate this model).",
        "type": "object",
        "required": ["coin_identifier", "coin_action"],
        "properties": {
          "coin_identifier": {
            "$ref": "#/components/schemas/CoinIdentifier"
          },
          "coin_action": {
            "$ref": "#/components/schemas/CoinAction"
          }
        }
      },
...
      "BlockTransactionRequest": {
        "description": "A BlockTransactionRequest is used to fetch a Transaction included in a block that is not returned in a BlockResponse.",
        "type": "object",
        "required": ["network_identifier", "block_identifier", "transaction_identifier"],
        "properties": {
          "network_identifier": {
            "$ref": "#/components/schemas/NetworkIdentifier"
          },
          "block_identifier": {
            "$ref": "#/components/schemas/BlockIdentifier"
          },
          "transaction_identifier": {
            "$ref": "#/components/schemas/TransactionIdentifier"
          }
        }
      },
      "BlockTransactionResponse": {
        "description": "A BlockTransactionResponse contains information about a block transaction.",
        "type": "object",
        "required": ["transaction"],
        "properties": {
          "transaction": {
            "$ref": "#/components/schemas/Transaction"
          }
        }
      },
...
      "Error": {
        "description": "Instead of utilizing HTTP status codes to describe node errors (which often do not have a good analog), rich errors are returned using this object. Both the code and message fields can be individually used to correctly identify an error. Implementations MUST use unique values for both fields.",
        "type": "object",
        "required": ["code", "message", "retriable"],
        "properties": {
          "code": {
            "description": "Code is a network-specific error code. If desired, this code can be equivalent to an HTTP status code.",
            "type": "integer",
            "format": "int32",
            "minimum": 0,
            "example": 12
          },
          "message": {
            "description": "Message is a network-specific error message. The message MUST NOT change for a given code. In particular, this means that any contextual information should be included in the details field.",
            "type": "string",
            "example": "Invalid account format"
          },
          "retriable": {
            "description": "An error is retriable if the same request may succeed if submitted again.",
            "type": "boolean"
          },
          "details": {
            "type": "object",
            "required": ["message"],
            "properties": {
              "message": { "type": "string" }
            },
            "description": "Often times it is useful to return context specific to the request that caused the error (i.e. a sample of the stack trace or impacted account) in addition to the standard error message.",
            "example": {
              "address": "0x1dcc4de8dec75d7aab85b567b6",
              "error": "not base64"
            }
          }
        }
      }
    }
  }
}